Moved testcases connecting to a HMC to end2end testcases; Other test cleanup
Details:.

* Deleted function test test_info_error_no_host() because that test is already
  contained in the testcases of end2end test test_session_logon().

* Moved function test test_info_error_no_conn() to the end2end test
  test_session_logon() as a new testcase, because it attempts to connect when
  a ~/.zhmc_sessions.yml file exists.

* Moved function test test_option_errorformat() to the end2end tests into
  a new end2end test module test_options.py, because it relies on a connection
  error when connecting to an unknown HMC host.

* Renamed the following function tests in test_options.py:
  - test_global_help() -> test_option_help()
  - test_global_version() -> test_option_version()

* Moved the following option-related function tests to test_options.py:
  - test_option_outputformat_table() -> test_option_format_table_props()
  - test_option_outputformat_json() -> test_option_format_json_props()
  - test_option_log()
  - test_option_logdest()
